Выполнил задание 5.1
